ml/requirements.txt
# Core
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=8.0.0
PyYAML>=5.4.0

# ML
torch>=1.9.0
torchvision>=0.10.0
ultralytics>=8.0.0

# Utilities
matplotlib>=3.3.0
seaborn>=0.11.0
tqdm>=4.60.0